requirements.txt
colorama==0.4.6
pillow==10.3.0
numpy==1.26.0